Add Mongoid::Criteria::Scrollable::Fields to reduce duplication
