+++ lib/mongo/scrollable.rb
@@ -1,5 +1,7 @@
 module Mongo
   module Scrollable
+    include Mongoid::Criteria::Scrollable::Fields
+
     def scroll(cursor = nil, options = nil, &_block)
       view = self
       # we don't support scrolling over a view with multiple fields
@@ -28,19 +30,6 @@ module Mongo
       else
         view
       end
-    end
-
-    private
-
-    def raise_mismatched_sort_fields_error!(cursor, criteria_cursor_options)
-      diff = cursor.sort_options.reject { |k, v| criteria_cursor_options[k] == v }
-      raise Mongoid::Scroll::Errors::MismatchedSortFieldsError.new(diff: diff)
-    end
-
-    def different_sort_fields?(cursor, criteria_cursor_options)
-      criteria_cursor_options[:field_type] = criteria_cursor_options[:field_type].to_s
-      criteria_cursor_options[:field_name] = criteria_cursor_options[:field_name].to_s
-      criteria_cursor_options != cursor.sort_options
     end
   end
 end
