+++ lib/mongoid-scroll.rb
@@ -7,6 +7,7 @@ require 'mongoid-compatibility'
 require 'mongoid/scroll/version'
 require 'mongoid/scroll/errors'
 require 'mongoid/scroll/cursor'
+require 'mongoid/criteria/scrollable/fields'
 require 'moped/scrollable' if Object.const_defined?(:Moped)
 require 'mongo/scrollable' if Object.const_defined?(:Mongo)
 require 'mongoid/criteria/scrollable'
